Explode crag table actions into separate components.
